backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
requests
radon
httpx
pyjwt
cryptography
pydriller
langchain
langchain-groq
